fix(ui): address CodeRabbit review — decimal input, slider, accessibility, validation
- PercentInput: use local state for intermediate typing ("2.", ".5")
- Slider: handle single numeric value prop (renders 1 thumb not 2)
- FormField: trim whitespace before Number() to prevent silent 0
- scenarioInputs: add .int() constraint to yearsIntoLoan
- Tabs: forward orientation prop to primitive for keyboard navigation
- AboutYouSection: add aria-describedby on filing status select
- CurrencyInput: document USD-only scope
- Badge: use ?? fallback to handle null variant

+++ src/schemas/scenarioInputs.ts
@@ -85,7 +85,7 @@ const currentHomeInputsSchema = z.object({
   /** Original loan term in years. */
   originalLoanTermYears: z.number().int().min(1).max(40),
   /** Number of years already into the loan. */
-  yearsIntoLoan: z.number().min(0).max(40),
+  yearsIntoLoan: z.number().int().min(0).max(40),
   /** Annual property tax rate as a fraction of assessed value (e.g. 0.0215). */
   annualPropertyTaxRate: rate(),
   /** Annual homeowners insurance premium in dollars. */
